test_requirements.txt
numpy>=1.10.0
pandas
Biopython
scipy
blosum
lmdb
pytest
